Feature/share style

See merge request treasurenft/treasure_nft_project!39

+++ lib/models/http/parameter/check_reserve_deposit.dart
@@ -21,7 +21,7 @@ class CheckReserveDeposit {
   int deposit;
   double reserveBalance;
   int tradingTime;
-  int reward;
+  double reward;
 
   factory CheckReserveDeposit.fromJson(Map<String, dynamic> json) =>
       CheckReserveDeposit(
